Extract OngoingStubbing to other class
